requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
plotly==5.18.0
openai==1.10.0
python-dotenv==1.0.0
openpyxl==3.1.2
